Refactor: Improve error handling
All handlers are currently "infallible". Removing this type argument from the result type of the service handlers sends ripples through the code: The warp handlers have to return 'Rejection' types which are incompatible with anyhow and the 'map_to_json' function no longer works with 'Infallible' errors.

It apparently also isn't possible to do error handling (as in determining the status code) directly within the handler. At least not without accepting that success- and error values have exactly the same type.

The solution seems to be adding custom error types and having them implement 'warp::reject::Reject'. A custom 'recover' function can then be registered after the handler to translate the errors into HTTP responses. This commit contains an example for how to set this up.