Jenkins: Fix build of debian package
The Jenkins worker is not Debian-based, so doesn't have the tools for building Debian packages installed. The solution is to run the Debian build script from within the builder container and extract the package from the pipeline instead of only extracting the binary and building the package in the pipeline.