Use slot time of last finalized block for 'current_block_timestamp' (#34)
The existing solution defaulted to -1 when the 'last_finalized_time' of 'consensus status' was None. This happens if the node wasn't online when the block was finalized and breaks the Rosetta CLI.

The slot time of the last finalized block, which is objectively agreed upon across all nodes, is now used instead.