Upgrade npm dependency
The previous run caused all kinds of updates to 'package-lock.json' anyway, so we might as well take it one step further. The bump doesn't change any generated output.

It is quite a mystery what all these dependencies are used for in a project that just downloads and runs a jar though...