Bump Rosetta spec (#49)
Bumping from v1.4.12 to v1.4.15. The spec authors forgot to bump the version field of the spec, so the generated files erroneously identify themselves as v1.4.13.

Installing/running the generator caused all kinds of changes to 'package-lock.json', so decided that we might as well take it one step further and upgrade the dependency. This bump doesn't cause any changes to the generated output. It is quite a mystery what all these dependencies are used for in a project that just downloads and runs a jar though...