Example tools: Add support for providing memo as string (#62)
The memo strings are CBOR encoded by the tool.

It was necessary to upgrade `transfer-client-direct` to using gRPC APIv2 for it to work with the Node. Also, the types used for amount to send and format of the key file was aligned between the two transfer tools. The new file format includes the account address, so we use that instead of passing it in the option `--sender`.

The changes have been fully verified by sending transfers on testnet (all combinations of the memo options).